Move function `SignTransaction` to package `signer`
