+++ transaction/signer/secp256k1_blake160_sighash_all.go
@@ -2,7 +2,10 @@ package signer
 
 import (
 	"bytes"
+	"encoding/binary"
 	"errors"
+	"fmt"
+	"github.com/nervosnetwork/ckb-sdk-go/crypto"
 	"github.com/nervosnetwork/ckb-sdk-go/crypto/blake2b"
 	"github.com/nervosnetwork/ckb-sdk-go/crypto/secp256k1"
 	"github.com/nervosnetwork/ckb-sdk-go/transaction"
@@ -27,7 +30,7 @@ func (s *Secp256k1Blake160SighashAllSigner) SignTransaction(tx *types.Transactio
 	}
 	if matched {
 		i0 := group.InputIndices[0]
-		signature, err := transaction.SignTransaction(tx, uint32ArrayToIntArray(group.InputIndices), tx.Witnesses[i0], ctx.Key)
+		signature, err := SignTransaction(tx, uint32ArrayToIntArray(group.InputIndices), tx.Witnesses[i0], ctx.Key)
 		if err != nil {
 			return false, err
 		}
@@ -49,4 +52,45 @@ func IsSingleSigMatched(key *secp256k1.Secp256k1Key, scriptArgs []byte) (bool, e
 	}
 	hash := blake2b.Blake160(key.PubKey())
 	return bytes.Equal(scriptArgs, hash), nil
+}
+
+// SignTransaction signs transaction with index group and witness placeholder in secp256k1_blake160_sighash_all way
+func SignTransaction(tx *types.Transaction, group []int, witnessPlaceholder []byte, key crypto.Key) ([]byte, error) {
+	inputsLen := len(tx.Inputs)
+	for i := 0; i < len(group); i++ {
+		if i > 0 && group[i] <= group[i-1] {
+			return nil, fmt.Errorf("group index is not in ascending order")
+		}
+		if group[i] > inputsLen {
+			return nil, fmt.Errorf("group index %d is greater than input bytesLen %d", group[i], inputsLen)
+		}
+	}
+	txHash := tx.ComputeHash()
+	msg := txHash.Bytes()
+	bytesLen := make([]byte, 8)
+	binary.LittleEndian.PutUint64(bytesLen, uint64(len(witnessPlaceholder)))
+	msg = append(msg, bytesLen...)
+	msg = append(msg, witnessPlaceholder...)
+
+	var indexes []int
+	for i := 1; i < len(group); i++ {
+		indexes = append(indexes, group[i])
+	}
+	for i := inputsLen; i < len(tx.Witnesses); i++ {
+		indexes = append(indexes, i)
+	}
+	for _, i := range indexes {
+		bytes := tx.Witnesses[i]
+		bytesLen := make([]byte, 8)
+		binary.LittleEndian.PutUint64(bytesLen, uint64(len(bytes)))
+		msg = append(msg, bytesLen...)
+		msg = append(msg, bytes...)
+	}
+
+	msgHash := blake2b.Blake256(msg)
+	signature, err := key.Sign(msgHash)
+	if err != nil {
+		return nil, err
+	}
+	return signature, nil
 }
