+++ transaction/signer/secp256k1_blake160_sighash_all_test.go
@@ -1,4 +1,4 @@
-package transaction
+package signer
 
 import (
 	"github.com/ethereum/go-ethereum/common"
